+++ R/issues_checks.R
@@ -18,23 +18,6 @@
 #'   issues <- issues_checks(meta = meta)
 #'   str(issues)
 issues_checks <- function(meta = meta_checks()) {
-  fields_check <- c(
-    "_linuxdevel",
-    "_macbinary",
-    "_winbinary",
-    "_status"
-  )
-  fields_info <- c(
-    "_buildurl"
-  )
-  fields <- c(fields_check, fields_info)
-  for (field in fields) {
-    meta[[field]][is.na(meta[[field]])] <- "src-failure"
-  }
-  success <- rep(TRUE, nrow(meta))
-  for (field in fields_check) {
-    success <- success & (meta[[field]] %in% c("success", "skipped"))
-  }
-  meta <- meta[!success,, drop = FALSE] # nolint
-  issues_list(meta[, c("package", fields)])
+  meta <- meta[lengths(meta$issues) > 0L,, drop = FALSE] # nolint
+  issues_list(meta[, c("package", "url", "issues")])
 }