+++ R/package.R
@@ -5,6 +5,7 @@
 #' @importFrom jsonlite parse_json read_json stream_in write_json
 #' @importFrom nanonext ncurl parse_url status_code
 #' @importFrom pkgsearch cran_package
+#' @importFrom rversions r_release r_versions
 #' @importFrom stats aggregate
 #' @importFrom utils available.packages compareVersion contrib.url unzip
 #' @importFrom vctrs vec_rbind vec_slice